backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
python-docx==1.1.0
spacy==3.7.2
transformers==4.36.0
torch==2.1.0
pydantic==2.5.0
numpy==1.24.3
scikit-learn==1.3.0
python-dateutil==2.8.2
nltk==3.8.1
pytesseract==0.3.10
Pillow==10.1.0
pandas==2.1.4
openpyxl==3.1.2